+++ failover-core/src/test/java/com/societegenerale/failover/core/scanner/FailoverScannerExceptionTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.societegenerale.failover.core.observable.scanner;
+package com.societegenerale.failover.core.scanner;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;