+++ failover-observable-micrometer/src/main/java/com/societegenerale/failover/observable/micrometer/health/FailoverHealthIndicator.java
@@ -16,7 +16,7 @@
 
 package com.societegenerale.failover.observable.micrometer.health;
 
-import com.societegenerale.failover.core.observable.scanner.FailoverScanner;
+import com.societegenerale.failover.core.scanner.FailoverScanner;
 import org.springframework.boot.health.contributor.Health;
 import org.springframework.boot.health.contributor.HealthIndicator;