+++ failover-spring-boot-autoconfigure/src/test/java/com/societegenerale/failover/configuration/BeanAssertions.java
@@ -29,7 +29,7 @@ import com.societegenerale.failover.core.payload.RecoveredPayloadHandler;
 import com.societegenerale.failover.core.observable.publisher.CompositeObservablePublisher;
 import com.societegenerale.failover.core.observable.FailoverObserver;
 import com.societegenerale.failover.core.observable.publisher.ObservablePublisher;
-import com.societegenerale.failover.core.observable.scanner.FailoverScanner;
+import com.societegenerale.failover.core.scanner.FailoverScanner;
 import com.societegenerale.failover.core.store.FailoverStore;
 import org.springframework.context.ApplicationContext;